moved cfgs
